Comment section finish

+++ newt-frontend/src/app/services/auth/auth.interceptor.ts
@@ -29,7 +29,7 @@ export class AuthInterceptor implements HttpInterceptor {
       }
     }, (error: any) => {
       if (error instanceof HttpErrorResponse) {
-        if (error.error.errors.includes('token is expired')) {
+        if (error.error.errors && error.error.errors.includes('token is expired')) {
           localStorage.clear()
           window.location.reload()
         }
